Finally fix lint-staged
Using lint-staged script from
https://github.com/okonet/lint-staged/issues/522

The main issue was that the script was passing in package.json as an
affected file and therefore was linting all packages (since all packages
are affected by it). The trick was to filter it down to only files
within the projects.

This could come back to bite us later but whatever